day 05
got tripped by the trim() for a long time (the first spot is empty, but because I trimed the input, the first spot is no longer empty and the stacks fo crates are not correct anymore)
also, I have still problems with cloning arrays in JS, so the quick fix was to read the input twice, so we start part 2 with a clean input.
Very unclean code this time. But my answers are correct :)