day 11, part 2
I did not find the solution for myself. Had to look it up on the AoC reddit. I do not fully understand what we do here, but I know that if we don't use the mod, the numbers grow to Infinity quickly.
I tried to use BigInt() first, but that did not produce the correct result, and I'm not sure why.

I hate the problems where you need to know the One Math Trick to get the answer.